Refactor: Move all files to app directory
