--- template-packs/frontend/nuxt/base/app.vue
@@ -1,19 +1,0 @@
-<template>
-  <NuxtLayout>
-    <NuxtPage />
-  </NuxtLayout>
-</template>
-
-<script setup lang="ts">
-useHead({
-  titleTemplate: (titleChunk) => {
-    return titleChunk ? `${titleChunk} - {{PROJECT_NAME}}` : '{{PROJECT_NAME}}'
-  },
-})
-
-// Initialize auth state on app mount
-const { initAuth } = useAuth()
-onMounted(() => {
-  initAuth()
-})
-</script>
