+++ template-packs/frontend/nuxt/base/app/composables/useApi.ts
@@ -1,0 +1,23 @@
+export const useApi = () => {
+  const config = useRuntimeConfig()
+  const token = useCookie('auth-token')
+
+  const api = $fetch.create({
+    baseURL: config.public.apiBaseUrl,
+    onRequest({ options }) {
+      if (!token.value) return
+      const headers = new Headers(options.headers as HeadersInit)
+      headers.set('Authorization', `Bearer ${token.value}`)
+      options.headers = headers
+    },
+    onResponseError({ response }) {
+      if (response.status === 401) {
+        const { logout } = useAuth()
+        logout()
+      }
+    },
+  })
+
+  return api
+}
+
