+++ template-packs/frontend/nuxt/base/app/layouts/auth.vue
@@ -1,0 +1,6 @@
+<template>
+  <div class="min-h-screen bg-gray-50">
+    <slot />
+  </div>
+</template>
+
