+++ template-packs/frontend/nuxt/base/app/middleware/auth.ts
@@ -1,0 +1,8 @@
+export default defineNuxtRouteMiddleware((to) => {
+  const { isAuthenticated } = useAuth()
+
+  if (!isAuthenticated.value) {
+    return navigateTo(`/auth/login?redirect=${to.fullPath}`)
+  }
+})
+
