+++ template-packs/frontend/nuxt/base/app/middleware/guest.ts
@@ -1,0 +1,8 @@
+export default defineNuxtRouteMiddleware(() => {
+  const { isAuthenticated } = useAuth()
+
+  if (isAuthenticated.value) {
+    return navigateTo('/dashboard')
+  }
+})
+
