+++ template-packs/frontend/nuxt/base/app/pages/dashboard/settings.vue
@@ -1,0 +1,14 @@
+<template>
+  <div>
+    <h1 class="text-xl font-semibold mb-4">Settings</h1>
+    <p class="text-gray-500">Sample route.</p>
+  </div>
+</template>
+
+<script setup lang="ts">
+definePageMeta({
+  layout: 'dashboard',
+  middleware: 'auth',
+})
+</script>
+
