+++ template-packs/frontend/nuxt/base/dist
@@ -1,0 +1,1 @@
+/workspace/template-packs/frontend/nuxt/base/.output/public
