--- template-packs/frontend/nuxt/base/layouts/auth.vue
@@ -1,5 +1,0 @@
-<template>
-  <div class="min-h-screen bg-gray-50">
-    <slot />
-  </div>
-</template>
