--- template-packs/frontend/nuxt/base/layouts/default.vue
@@ -1,71 +1,0 @@
-<template>
-  <div class="min-h-screen bg-gray-50">
-    <!-- Navigation -->
-    <nav class="bg-white shadow">
-      <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-        <div class="flex justify-between h-16">
-          <div class="flex">
-            <div class="flex-shrink-0 flex items-center">
-              <NuxtLink to="/" class="text-xl font-bold text-indigo-600">
-                {{PROJECT_NAME}}
-              </NuxtLink>
-            </div>
-            <div class="hidden sm:ml-6 sm:flex sm:space-x-8">
-              <NuxtLink
-                to="/"
-                class="border-transparent text-gray-500 hover:border-gray-300 hover:text-gray-700 inline-flex items-center px-1 pt-1 border-b-2 text-sm font-medium"
-              >
-                Home
-              </NuxtLink>
-              <NuxtLink
-                v-if="isAuthenticated"
-                to="/dashboard"
-                class="border-transparent text-gray-500 hover:border-gray-300 hover:text-gray-700 inline-flex items-center px-1 pt-1 border-b-2 text-sm font-medium"
-              >
-                Dashboard
-              </NuxtLink>
-            </div>
-          </div>
-          
-          <div class="flex items-center space-x-4">
-            <template v-if="!isAuthenticated">
-              <NuxtLink
-                to="/auth/login"
-                class="text-gray-500 hover:text-gray-700 px-3 py-2 rounded-md text-sm font-medium"
-              >
-                Login
-              </NuxtLink>
-              <NuxtLink
-                to="/auth/register"
-                class="bg-indigo-600 hover:bg-indigo-700 text-white px-3 py-2 rounded-md text-sm font-medium"
-              >
-                Sign Up
-              </NuxtLink>
-            </template>
-            <template v-else>
-              <UserDropdown />
-            </template>
-          </div>
-        </div>
-      </div>
-    </nav>
-
-    <!-- Page Content -->
-    <main>
-      <slot />
-    </main>
-
-    <!-- Footer -->
-    <footer class="bg-white mt-auto">
-      <div class="max-w-7xl mx-auto py-6 px-4 sm:px-6 lg:px-8">
-        <p class="text-center text-sm text-gray-500">
-          © {{ new Date().getFullYear() }} {{PROJECT_NAME}}. All rights reserved.
-        </p>
-      </div>
-    </footer>
-  </div>
-</template>
-
-<script setup lang="ts">
-const { isAuthenticated } = useAuth()
-</script>
