--- template-packs/frontend/nuxt/base/middleware/guest.ts
@@ -1,7 +1,0 @@
-export default defineNuxtRouteMiddleware((to, from) => {
-  const { isAuthenticated } = useAuth()
-
-  if (isAuthenticated.value) {
-    return navigateTo('/dashboard')
-  }
-})
