+++ template-packs/frontend/nuxt/base/package.json
@@ -24,18 +24,22 @@
     "@nuxtjs/tailwindcss": "latest",
     "@pinia/nuxt": "latest",
     "@vueuse/nuxt": "latest",
+    "pinia": "latest",
     "nuxt": "^4.1.0",
     "vue": "^3.5.0",
-    "vue-router": "^4.0.0"
+    "vue-router": "^4.4.0"
   },
   "devDependencies": {
     "@nuxt/devtools": "latest",
-    "@nuxt/test-utils": "latest",
     "@nuxtjs/eslint-config-typescript": "^12.1.0",
     "@typescript-eslint/eslint-plugin": "^6.14.0",
     "@typescript-eslint/parser": "^6.14.0",
     "@vitest/coverage-v8": "^1.0.0",
     "@vue/test-utils": "^2.4.3",
+    "@types/node": "^20.11.28",
+    "@tailwindcss/forms": "^0.5.7",
+    "@tailwindcss/typography": "^0.5.10",
+    "@tailwindcss/aspect-ratio": "^0.4.2",
     "autoprefixer": "^10.4.16",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
