+++ template-packs/frontend/nuxt/base/tailwind.config.js
@@ -1,12 +1,8 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
-    "./components/**/*.{js,vue,ts}",
-    "./layouts/**/*.vue",
-    "./pages/**/*.vue",
-    "./plugins/**/*.{js,ts}",
-    "./nuxt.config.{js,ts}",
-    "./app.vue"
+    "./app/**/*.{js,vue,ts}",
+    "./nuxt.config.{js,ts}"
   ],
   theme: {
     extend: {
