+++ template-packs/frontend/nuxt/base/tsconfig.json
@@ -6,8 +6,6 @@
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
     "types": [
-      "@nuxt/types",
-      "@pinia/nuxt",
       "@types/node"
     ]
   }
